requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
distlib==0.3.0
dj-static==0.0.6
Django==3.0.5
django-allauth==0.41.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.0
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.1
pipenv==2018.11.26
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.17
virtualenv-clone==0.5.4
zipp==3.1.0
